Product backend integration done

+++ src/app/main/products/product-settings/product-settings.component.ts
@@ -1,11 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
-import { subscribeOn } from 'rxjs';
 import { ProductListService } from 'src/app/services/product-list.service';
 import { AddChangeDialogComponent } from './add-change-dialog/add-change-dialog.component';
 import { ProductList } from '../productList';
 import { MatTableDataSource } from '@angular/material/table';
 import { AccountService } from 'src/app/services/account.service';
+import { AlertifyService } from 'src/app/services/alertify.service';
+
+declare let alertify: any;
 
 @Component({
   selector: 'app-product-settings',
@@ -17,23 +19,30 @@ export class ProductSettingsComponent implements OnInit {
   dataSource: any;
   tableData: any;
   selectedProductsId: number[] = [];
-  some!: boolean
+  some!: boolean;
   addProductPermission!: boolean;
   editProductPermission!: boolean;
   deleteProductPermission!: boolean;
 
-
   constructor(
     private productListService: ProductListService,
     private accountService: AccountService,
-    public dialog: MatDialog
-  ) { }
+    public dialog: MatDialog,
+    private alertifyService:AlertifyService
+  ) {}
 
   ngOnInit(): void {
-    const accountPermissions = this.accountService.getAccountPermissions()[0].permissions
-    this.addProductPermission = accountPermissions.some((p: any) => p == "addProduct");
-    this.editProductPermission = accountPermissions.some((p: any) => p == "changeProduct");
-    this.deleteProductPermission = accountPermissions.some((p: any) => p == "removeProduct");
+    const accountPermissions =
+      this.accountService.getAccountPermissions()[0].permissions;
+    this.addProductPermission = accountPermissions.some(
+      (p: any) => p == 'addProduct'
+    );
+    this.editProductPermission = accountPermissions.some(
+      (p: any) => p == 'changeProduct'
+    );
+    this.deleteProductPermission = accountPermissions.some(
+      (p: any) => p == 'removeProduct'
+    );
     this.refresh();
   }
 
@@ -42,7 +51,6 @@ export class ProductSettingsComponent implements OnInit {
       this.dataSource = data;
       this.tableData = new MatTableDataSource(this.dataSource);
     });
-
   }
 
   applyFilter(event: Event) {
@@ -53,23 +61,23 @@ export class ProductSettingsComponent implements OnInit {
   edit(id: number) {
     if (this.editProductPermission) {
       const clickedRowData = this.dataSource.filter((d: any) => {
-        return d.id == id
-      })
-      const objectClickedRowData = clickedRowData[0]
+        return d.id == id;
+      });
+      const objectClickedRowData = clickedRowData[0];
       const dialogRef = this.dialog.open(AddChangeDialogComponent, {
-        width: "550px",
-        height: "900px",
-        data: objectClickedRowData
-      })
+        width: '550px',
+        height: '900px',
+        data: objectClickedRowData,
+      });
 
-      dialogRef.afterClosed().subscribe(data => {
+      dialogRef.afterClosed().subscribe((data) => {
         if (data) {
-          data.id = id
+          data.id = id;
           this.productListService.putProductList(data, id).subscribe(() => {
             this.refresh();
           });
         }
-      })
+      });
     }
   }
 
@@ -77,21 +85,25 @@ export class ProductSettingsComponent implements OnInit {
     if (e.checked) {
       this.selectedProductsId.push(id);
     } else {
-      const index = this.selectedProductsId.indexOf(id)
+      const index = this.selectedProductsId.indexOf(id);
       if (index > -1) {
         this.selectedProductsId.splice(index, 1);
       }
     }
-    console.log(this.selectedProductsId)
   }
 
   removeSelectedRows() {
-    this.selectedProductsId.map((id: any) => {
-      this.productListService.deleteProductList(id).subscribe(() => {
-        this.refresh();
-      })
-    })
+    alertify.confirm("DELETE!!!",'Are you sure you want to delete selected rows?', () => {
+      this.selectedProductsId.map((id: any) => {
+        this.productListService.deleteProductList(id).subscribe(() => {
+          this.refresh();
+          this.alertifyService.success("Successfuly deleted")
+        });
+      });
+    }, () => {
+    });
   }
+   
 
   addProduct() {
     const newProduct: ProductList = {
@@ -100,21 +112,21 @@ export class ProductSettingsComponent implements OnInit {
       category: '',
       url: '',
       price: 0,
-      description: ''
-    }
+      description: '',
+    };
     const dialogRef = this.dialog.open(AddChangeDialogComponent, {
-      width: "550px",
-      height: "900px",
-      data: newProduct
-    })
+      width: '550px',
+      height: '900px',
+      data: newProduct,
+    });
 
-    dialogRef.afterClosed().subscribe(data => {
+    dialogRef.afterClosed().subscribe((data) => {
       if (data) {
-        data.id = 0
+        data.id = 0;
         this.productListService.postProductList(data).subscribe(() => {
           this.refresh();
-        })
+        });
       }
-    })
+    });
   }
 }
